+++ src/entities/clothes/index.ts
@@ -1,7 +1,6 @@
 export {
   deleteLookbookItem,
   getAllLookbookListByWeather,
-  getMemberLookbook,
   saveLookbook,
 } from './api/lookbook';
