+++ src/widgets/clothes/index.ts
@@ -1,5 +1,6 @@
 export { getRecommendClothes } from './api/recommend';
 
-export type { RecommendClothesDto } from './model/types';
+export type { RecommendClothesDto, MemberLookbookDto } from './model/types';
 
 export { RecommendClothesList } from './ui/RecommendClothesList/RecommendClothesList';
+export { HomeLookbookList } from './ui/HomeLookbookList/HomeLookbookList';
